Move MapNeighborhood to neighborhood package
More neighborhood related ops will be added in the future and this will
hopefully avoid the map package getting cluttered.
